fix: update TS types for `PubSub#subscription` to match 0.23+'s API
As of 0.23 and later, `PubSub#subscription` and `Topic#subscription` take an options object with `streamingOptions` instead of `maxConnections`. Also, the methods that take this options object never use `gaxOpts` nor `autoPaginate` -- these fields are unused in these code paths.

This commit updates the TS types so that the options object passed to `PubSub#subscription` and `Topic#subscription` (a) has the new `streamingOptions` field and (b) does not have the unused `gaxOpts` and `autoPaginate` fields. As part of this commit, some TS `as` type assertions are no longer needed.

There is one small breaking change in this commit -- the options passed to `Topic#subscription` are no longer mutated. Instead, a new options object (of a different type) is created. Since the options are no longer mutated, a small change to one of the unit tests was necessary.

Test plan: TS builds and all tests pass