feat: Enable server side flow control by default with the option to turn it off
This change enables sending flow control settings automatically to the server.
If flowControl.maxMessages > 0 or flowControl.maxBytes > 0, flow control will
be enforced at the server side (in addition to the client side).

This behavior is enabled by default and SubscriberOptions.useLegacyFlowControl can be used for users who would
like to opt-out of this feature in case they encounter issues with server side
flow control.